fix: address code review findings for #475–#477 (#484)
C1: Wire closeStory() after runPipeline() in iteration-runner.ts so all
    non-terminal sessions for a story are marked COMPLETED at end of each
    iteration. Update run-setup.ts comments to clarify Phase 5.5 plan for
    sweepOrphans() and onShutdown.

H1: Fix role mismatch in bindHandle — context.ts now creates descriptor
    with role "implementer" to match the sessionRole used in execution.ts,
    ensuring deriveSessionName(descriptor) produces the same ACP session
    handle as the physical session.

H2: Fix misleading comment in adapter.ts finally block — correctly states
    that on failure (any) the session is kept open for retry context.

M1: Fix barrel imports — AdapterFailure and ToolDescriptor now imported
    from "../../context/engine" barrel in adapter.ts and types.ts.

M2: Split adapter-session.test.ts (633→346 lines) — extract interaction
    bridge/context-pull tests to adapter-interaction.test.ts and Phase 1
    plumbing tests to adapter-phase1.test.ts (adds deriveSessionName tests).

M3: Split manager.test.ts (407→345 lines) — extract bindHandle() tests
    to manager-bind-handle.test.ts.

M4: Add PAUSED/RESUMING/CLOSING state coverage to closeStory() tests in
    manager-lifecycle.test.ts (4 new tests, 1 resume-after-close variant).

L1: Update adapter.ts file header — replace deleted sidecar reference with
    SessionManager disk-backed descriptor path.

L2: Add debug log to SessionManager.resume() on successful match for
    operational visibility in parallel runs.
